clean up server infra methods. add tests for nutrition and ingredient repositories
