+++ backend/src/installRoutes.ts
@@ -1,0 +1,5 @@
+import {Application} from 'express';
+
+export function installRoutes(app: Application) {
+  // TODO: Implement routes
+}
