+++ backend/src/recipes/recipesRepository.ts
@@ -12,6 +12,7 @@ export async function getRecipes(): Promise<{}> {
     client.release();
   }
 }
+
 export async function createRecipe(recipeData: {
   name: string;
   ingredients: {id: number; quantity: number; unit: string}[];
