+++ backend/src/routes/recipesRouter.ts
@@ -1,0 +1,11 @@
+import {Router} from 'express';
+
+export const recipesRouter = Router();
+
+recipesRouter.post('/', async (req, res) => {
+  // TODO: Implement post recipe
+});
+
+recipesRouter.get('/', async (req, res) => {
+  // TODO: Implement get recipe
+});
